requirements.txt
faiss-cpu>=1.8.0
numpy>=1.26.0
openai>=1.35.0
pypdf>=4.2.0
sentence-transformers>=3.0.0
